b2b/requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.4
psycopg2-binary==2.8.4
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.6.0
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
